scala: Run steps directly from jar file
The Scala build will create a mal.jar file with all the steps' classes
in it.  A `run` wrapper script will run that jar with a specific step
main class according to the `$STEP` environment variable.  Now the jar
file is created only once (when testing step0) and then re-used for
testing the next steps, which should speed up the build.

Moreover, sbt now prepends shell script header to the jar file, so no
need to implement this in Makefile (for the `dist` target).