Fix bug in read_str for long strings
Symbols are read by the tokenizer into RSI, but RSI
was overwritten when a new string chunk was fetched.
If the reader happened to be reading a symbol as the
end of an Array chunk was reached, then the symbol
would be changed to part of the input string.

All step 8 tests now pass.