perl: Use goto &NAME for tail-call optimisation.
This is rather contrary to the guide, which expects a loop in EVAL, but
perl is unusual in having native support for (manual) tail-call
optimisation and it would seem a shame not to take advantage of this.

This leaves macros broken since all functions are now implemented as
core functions, and we don't yet have support for turning them into
macros.