runtest: Decode UTF-8 from implementation correctly
Previously, non-ASCII bytes from the implementation would lead to
errors.  Under Python 2 this was because the input was assumed to be
ASCII, while under Python 3 it was assumed to be UTF-8 but decoded by
calling "decode" on one byte at a time.  Now each Runner has an
IncrementalDecoder that deals with this properly, since Python itself
refuses to construct an unbuffered text stream.