bcpl: Extend garbage-collection to user-defined functions
This adds a field, fun_ntracked, that indicates how many of the fields
of a user-defined function are tracked pointers.  This gets set when the
function is created and inspected by gc_mark(), but can be otherwise
ignored.  At the moment it's always 0 (for core functions) or 3 (for
functions defined using fn*).