bcpl: Set up exception handler earlier
It's nice to catch exceptions thrown by the startup calls to rep(), for
those cases where you've really broken things.  From step 6 onwards we
already have an exception handler that simply prints a message and
exits, so enable that before calling rep().  If we go into the REPL
(i.e. no file is supplied), it then gets replaced by an exception
handler that returns to the REPL.