bcpl: Garbage-collect less often for speed.
A mark-and-sweep garbage-collector has an overhead each run of the
number of live objects, so running too often is expensive.  The
garbage-collector now only runs if the number of objects allocated since
the last run exceeds the number that were live at the end of the last
run.  This is a fairly common heuristic and improves performance
markedly.

Consequently, remove the test timeout adjustment.  It's more than fast
enough now.