Fix formatting of note about `throw` when defining `cond` at the end of Step 8
This change removes one level of indentation on the note about using `throw` (which is not yet defined) when defining the `cond` macro at the end of Step 8. The Markdown engine on Github cannot start an indented list following a block of code, so it treats the bullet as 4-space indented code, causing the bullet point to be formatted as ugly pre-formatted text.