Temporary Commit
It's still a mess and unusable now. But I know what to do now.

When an AST (A) gets compiled, the end result is a code (C) to be
executed. Execution of C should return a *closure* (not a function,
which I've been doing, which is wrong), that holds a slot telling the
function where to look for literals (e.g. scalars, functions, objects..
etc). So far I've been trying to save those literals in a global
constant _consts. It was why my code failed so far because some compiled
ASTs may be alive even after the next compilation. As a consequence, as
they are reading from the global _consts, they read the wrong data.