runtest.py: if --no-pty, do not expect echo'd input
If --no-pty is set, then don't expect the input forms to be echo'd to
stdout. This should help when running in WSL on Windows and calling into
a regular Windows executable.

This should mostly work if the input/test forms are echo'd, however, it
could potentially cause issues for step0 tests where it may appear that
things are successful because this may match the input form rather than
the output. So we limit this to --no-pty which is probably only going to
be used for the Windows WSL case in the future.