common-lisp: (list? nil) should be false (add test)
In common-lisp, self-hosted test4 was failing and it appeared that nil
and empty list () were being conflated. The root cause is that in
common-lisp the form `(list? nil)` was returning true. Fix this in the
core `list?` function.

There was no test for this case so add one to step4 tests.