basic: aggressively reduce diff between steps
Copy structural changes from high steps to low steps (mostly, split RE
out of REP) in order to reduce the diffs (mostly, 3 to 4 which is
already too large).

On the contrary, remove similar backports that only increase the
complexity of first steps without reducing the diffs:
 * push the environment in the stack in step2 (it does not change yet)
 * release env if not on top of the stack in step3 (not TCO yet)
 * ignoring metadata in steps4-9 (no metadata yet)

Remove obsolete debug assignments from the start of DO_FUNCTION in
step2.

Reorder the load-file logic in steps6-A in order to reduce the diff
between steps 9 and A.

Aggressively remove differences in spaces and comments.