yamlscript: WIP - A new implementation
Passing steps 0-3 so far, with no skipped tests.

To test, run `make -C impls/yamlscript test`.

The Makefile will auto-install the ys binary under the
`impls/yamlscript/.cache/.local/bin/` directory.

In other words, it should "just work" with no prerequisites.

YAMLScript (aka YS) was created in 2023 after I finished a mal
implementation in Perl.
Without the mal project, YS would not exist!

See: https://yamlscript.org

The code in this implementaion was ported almost directly from the
impls/mal "code" (which cannot run without a hosting implementation).

This is to say that the YS mal code (like the MAL mal code) reads much
like pseudocode, the difference being that the YS code actually runs!
And it can host the MAL mal code (of course).

This works out so well, because YS transpiles to Clojure.
It is literally a syntax dialect of Clojure (without a JVM, thanks to
the  GraalVM native-image compiler).