runtest: show test results in timed out case.
A timed out test means that the tests can't proceed (REPL likely hung).
However, we still want to get a test summary for that case. When
a timeout happens, instead of throw an exception and immediately
existing, the test file continues to be read but no more tests are
executed (similar to if a regular test failure happens without
--continue-on-fail).